+++ src/main/java/com/ssafy/gilbut/domain/board/service/BoardService.java
@@ -11,11 +11,11 @@ import org.springframework.data.domain.Pageable;
 public interface BoardService {
 
 	SimplePageResultDTO listArticle(Pageable pageable);
-	DetailResultDTO getArticle(Integer boardId);
-	DetailResultDTO getModifyArticle(String accessToken, Integer boardId);
+	DetailResultDTO getArticle(Long boardId);
+	DetailResultDTO getModifyArticle(String accessToken, Long boardId);
 	void writeArticle(String accessToken, CreateDTO boardDto);
-	void modifyArticle(String accessToken, Integer boardId, UpdateDTO boardDto);
-	void updateHit(Integer boardId);
+	void modifyArticle(String accessToken, Long boardId, UpdateDTO boardDto);
+	void updateHit(Long boardId);
 
-	void deleteArticle(String accessToken, Integer boardId);
+	void deleteArticle(String accessToken, Long boardId);
 }
