+++ src/main/java/com/ssafy/gilbut/domain/plan/mapper/PlanMapper.java
@@ -14,24 +14,24 @@ public interface PlanMapper {
 
     void insertPlans(@Param("dto") List<PlanRequest.CreateDTO> dto);
 
-    Optional<Plan> findPlanById(@Param("planId") Integer planId);
+    Optional<Plan> findPlanById(@Param("planId") Long planId);
 
     List<Plan> findPlansByUserId(
-            @Param("userId") Integer userId,
+            @Param("userId") Long userId,
             @Param("pageable") Pageable pageable
     );
 
     List<Plan> findPlansByPlanIds(
-            @Param("ids") List<Integer> ids
+            @Param("ids") List<Long> ids
     );
 
-    int countPlansByUserId(@Param("userId") Integer userId);
+    int countPlansByUserId(@Param("userId") Long userId);
 
     void updatePlans(
-            @Param("planId") Integer planId,
+            @Param("planId") Long planId,
             @Param("dto") List<PlanRequest.UpdateDTO> dto
     );
 
 
-    void deletePlanByTravelId(@Param("travelId") Integer travelId);
+    void deletePlanByTravelId(@Param("travelId") Long travelId);
 }
